Refactors authentication flow and splash screen
Updates the authentication flow to check for existing users upon app launch.  The splash screen now redirects to the home screen if a user is logged in, and to the authentication screen otherwise.  Improves the `AuthCubit` to manage user state more effectively and adds a method to retrieve the current user.  Removes the in-app game from the home screen.